Change client directory structure
