+++ client/src/components/component.ts
@@ -1,0 +1,7 @@
+export default abstract class Component {
+	protected dom: HTMLElement | null = null;
+	getDom(): HTMLElement {
+		if (this.dom === null) throw new Error('Initialize dom');
+		return this.dom;
+	}
+}
