+++ client/src/components/content/history-content/index.ts
@@ -1,0 +1,3 @@
+import HistoryView from './history-view';
+
+export default HistoryView;
